attempt to refactor fizzbuzz
