+++ build.sbt
@@ -1,0 +1,2 @@
+scalaVersion := "2.13.8"
+libraryDependencies += "org.scalatest" %% "scalatest" % "3.2.9" % Test
